requirements.txt
pytest==6.2.5
selenium==4.1.0
allure-pytest==2.9.45
allure-python-commons==2.9.45